refactor: accept middleware executor to run underlying functions
BREAKING CHANGE: After this change, the "run" method does not accept the parameters
to pass to the middleware, instead it accepts a executor function. The executor function
receives the middleware to execute, along with the next function.

This allows the middleware pipeline to be flexible with the data the middleware receives